start of data editor tool
tackling unit definitions and found that simple property editors for assets were insufficient for the task.
data editor tool will allow easy creation and editing of mods and their data
so far just working on getting everything i need for the navigation tree. it can list mods and nestable 'trait collections' so far.
each of these tree items will be able to push editor widgets to the main pane of the window